+++ CHANGELOG.md
@@ -3,7 +3,7 @@ English | [简体中文](./CHANGELOG_CN.md)
 #### V3.x.x (dev)
 
 - [FEATURE] Add `vConsole.showSwitch()` and `vConsole.hideSwitch()` methods, see [Public Properties & Methods](./doc/public_properties_methods.md).
-- [FEATURE] Add `onReady` callback function to `vConsole.option`.
+- [FEATURE] Add `onReady` and `onClearLog` callback function to `vConsole.option`.
 
 
 #### V3.0.0 (2017-09-27)